+++ java/src/com/protobytes/buffers/core/BuffFloat.java
@@ -3,20 +3,20 @@ package com.protobytes.buffers.core;
 import com.protobytes.buffers.BuffType;
 
 /**
- * 4个字节，单精度（32 位）浮点数。
+ * A single-precision (32-bit) floating-point number.
  * 
  * @author Agua.L
  *
  */
 public class BuffFloat extends BuffBase {
-	private Float value = 0.0f;
+	private Double value = 0.0d;
 
 	public Object getValue() {
 		return this.value;
 	}
 
 	public void setValue(Object value) {
-		this.value = (Float) value;
+		this.value = (Double) value;
 	}
 
 	public int getType() {
